requirements.txt
alembic==1.14.0
astroid==3.3.6
azure-cognitiveservices-speech==1.41.1
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.0
azure-storage-file-share==12.20.0
azure-storage-queue==12.12.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==44.0.0
dill==0.3.9
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
lazy-object-proxy==1.10.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
msal==1.31.1
msal-extensions==1.2.0
platformdirs==4.3.6
portalocker==2.10.1
pycparser==2.22
pydub==0.25.1
PyJWT==2.10.1
pylint==3.3.2
PySocks==1.7.1
pywin32==308
requests==2.32.3
requests-file==2.1.0
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
tldextract==5.1.3
toml==0.10.2
tomlkit==0.13.2
typed_ast==1.5.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
zope.event==5.0
zope.interface==7.2
